requirements-dev.txt
-r src/backend/requirements.txt
ruff
mypy
types-requests
pre-commit
pip-tools
pip-compile-cross-platform
playwright
pytest
pytest-asyncio
pytest-cov
pytest-playwright
pytest-snapshot
locust
psycopg2
dotenv-azd
PyMuPDF
tqdm
pdfplumber